backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
psycopg[binary]==3.3.3
alembic==1.13.2
pydantic==2.9.2
pydantic-settings==2.5.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.10
email-validator==2.2.0
openai>=1.40.0
pymupdf>=1.24.0
python-docx>=1.1.0
pgvector>=0.3.0
jsonschema>=4.23.0
